Move _app.tsx to layout
